Extracted tabs factory from Main form
