--- Src/SwqlStudio/MainForm.ApplicationService.cs
@@ -1,67 +1,0 @@
-﻿using System.Windows.Forms;
-using SwqlStudio.Metadata;
-using SwqlStudio.Properties;
-using SwqlStudio.Subscriptions;
-using WeifenLuo.WinFormsUI.Docking;
-
-namespace SwqlStudio
-{
-    partial class MainForm : IApplicationService
-    {
-        public SubscriptionManager SubscriptionManager { get; } = new SubscriptionManager();
-
-        public void AddTextToEditor(string text, ConnectionInfo info)
-        {
-            if (info == null)
-                info = this.filesDock.ActiveConnectionInfo;
-
-            IMetadataProvider metadataProvider;
-            _metadataProviders.TryGetValue(info, out metadataProvider);
-
-            CreateQueryTab(info.Title, info, metadataProvider);
-            this.filesDock.ActiveQueryTab.QueryText = text;
-        }
-
-        public void OpenActivityMonitor(string title, ConnectionInfo info)
-        {
-            var activityMonitorTab = new ActivityMonitorTab { ConnectionInfo = info, Dock = DockStyle.Fill, ApplicationService = this };
-            AddNewTab(activityMonitorTab, title);
-            activityMonitorTab.Start();
-        }
-
-        public void OpenInvokeTab(string title, ConnectionInfo info, Verb verb)
-        {
-            var invokeVerbTab = new InvokeVerbTab { ConnectionInfo = info, Dock = DockStyle.Fill, ApplicationService = this, Verb = verb };
-            AddNewTab(invokeVerbTab, title);
-        }
-
-        /// <inheritdoc />
-        public void OpenCrudTab(CrudOperation operation, ConnectionInfo info, Entity entity)
-        {
-            string title = entity.FullName + " - " + operation;
-            var crudTab = new CrudTab(operation)
-            {
-                ConnectionInfo = info,
-                Dock = DockStyle.Fill,
-                ApplicationService = this,
-                Entity = entity
-            };
-
-            crudTab.CloseItself += (s, e) =>
-            {
-                this.filesDock.RemoveTab(crudTab.Parent as DockContent);
-            };
-
-            AddNewTab(crudTab, title);
-        }
-
-        private void AddNewTab(Control childControl, string title)
-        {
-            var dockContent = new DockContent();
-            dockContent.Icon = Resources.TextFile_16x;
-            dockContent.Controls.Add(childControl);
-            dockContent.Text = title;
-            dockContent.Show(this.filesDock, DockState.Document);
-        }
-    }
-}
