+++ Src/SwqlStudio/MainForm.cs
@@ -1,34 +1,29 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.ServiceModel;
-using System.ServiceModel.Security;
 using System.Windows.Forms;
-using ScintillaNET;
 using SolarWinds.InformationService.Contract2;
 using SolarWinds.InformationService.InformationServiceClient;
 using SwqlStudio.Metadata;
 using SwqlStudio.Properties;
 using SwqlStudio.Subscriptions;
-using WeifenLuo.WinFormsUI.Docking;
 
 namespace SwqlStudio
 {
     [ServiceBehavior(InstanceContextMode = InstanceContextMode.Single,
         ConcurrencyMode = ConcurrencyMode.Multiple,
         UseSynchronizationContext = false)]
-    public partial class MainForm : Form
+    internal partial class MainForm : Form, IApplicationService
     {
         private static readonly SolarWinds.Logging.Log log = new SolarWinds.Logging.Log();
-        
-        private readonly ServerList serverList = new ServerList();
-        private readonly Dictionary<ConnectionInfo, IMetadataProvider> _metadataProviders = new Dictionary<ConnectionInfo, IMetadataProvider>();
 
         public PropertyBag QueryParameters
         {
             get { return this.filesDock.QueryParameters; }
         }
+        
+        public SubscriptionManager SubscriptionManager { get; } = new SubscriptionManager();
+
 
         public MainForm()
         {
@@ -44,164 +39,30 @@ namespace SwqlStudio
 
         private void InitializeDockPanel()
         {
-            this.filesDock.ObjectExplorerImageList = this.ObjectExplorerImageList;
-            this.filesDock.ApplicationService = this;
+            this.filesDock.SetObjectExplorerImageList(this.ObjectExplorerImageList);
+            this.filesDock.SetAplicationService(this);
         }
 
         private void startTimer_Tick(object sender, EventArgs e)
         {
             startTimer.Enabled = false;
-            AddNewQueryTab();
+            this.filesDock.AddNewQueryTab();
         }
-
 
         #region Code related to File menu
 
         private void menuFileNew_Click(object sender, EventArgs e)
         {
-            AddNewQueryTab();
-        }
-
-        private void AddNewQueryTab()
-        {
-            using (NewConnection nc = new NewConnection())
-            {
-                if (nc.ShowDialog(this) != DialogResult.OK)
-                    return;
-
-                string msg = null;
-
-                try
-                {
-                    ConnectionInfo info;
-                    bool alreadyExists = false;
-                    alreadyExists = serverList.TryGet(nc.ConnectionInfo.ServerType, nc.ConnectionInfo.Server, nc.ConnectionInfo.UserName, out info);
-                    if (!alreadyExists)
-                    {
-                        info = nc.ConnectionInfo;
-                        info.Connect();
-                        serverList.Add(info);
-
-                        info.ConnectionClosed += (sender, args) => serverList.Remove(info);
-                    }
-
-                    if (!alreadyExists)
-                    {
-                        var provider = new SwisMetaDataProvider(info);
-                        this.filesDock.AddServer(provider, info);
-                        _metadataProviders[info] = provider;
-                    }
-
-                    CreateQueryTab(info.Title, info, _metadataProviders[info]);
-                }
-                catch (FaultException<InfoServiceFaultContract> ex)
-                {
-                    log.Error("Failed to connect", ex);
-                    msg = ex.Detail.Message;
-                }
-                catch (SecurityNegotiationException ex)
-                {
-                    log.Error("Failed to connect", ex);
-                    msg = ex.Message;
-                }
-                catch (FaultException ex)
-                {
-                    log.Error("Failed to connect", ex);
-                    msg = (ex.InnerException != null) ? ex.InnerException.Message : ex.Message;
-                }
-                catch (MessageSecurityException ex)
-                {
-                    log.Error("Failed to connect", ex);
-                    if (ex.InnerException != null && ex.InnerException is FaultException)
-                    {
-                        msg = (ex.InnerException as FaultException).Message;
-                    }
-                    else
-                    {
-                        msg = ex.Message;
-                    }
-                }
-                catch (Exception ex)
-                {
-                    log.Error("Failed to connect", ex);
-                    msg = ex.Message;
-                }
-
-                if (msg != null)
-                {
-                    msg = string.Format("Unable to connect to Information Service. {0}", msg);
-                    MessageBox.Show(this, msg, "Connection Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
-            }
-        }
-
-        private QueryTab CreateQueryTab(string title, ConnectionInfo info, IMetadataProvider provider)
-        {
-            var queryTab = new QueryTab { ConnectionInfo = info, Dock = DockStyle.Fill, ApplicationService = this };
-
-            queryTab.SetMetadataProvider(provider);
-
-            AddNewTab(queryTab, title);
-
-            info.ConnectionClosed += (sender, args) =>
-            {
-                this.filesDock.RemoveTab(queryTab.Parent as DockContent);
-            };
-
-            return queryTab;
+            this.filesDock.AddNewQueryTab();
         }
 
         private void menuFileOpen_Click(object sender, EventArgs e)
         {
             if (openFileDialog.ShowDialog() == DialogResult.OK)
-                // Try to open chosen file
-                OpenFiles(openFileDialog.FileNames);
+                this.filesDock.OpenFiles(openFileDialog.FileNames);
         }
 
-        private void OpenFiles(string[] fns)
-        {
-            var originalConnection = this.filesDock.ActiveConnectionInfo;
-            if (originalConnection == null)
-                return;
-
-            // Close default untitled document if it is still empty
-            this.filesDock.ColoseInitialDocument();
-
-            // Open file(s)
-            foreach (string fn in fns)
-            {
-                QueryTab queryTab = null;
-                try
-                {
-                    var connectionInfo = originalConnection.Copy();
-                    connectionInfo.Connect();
-
-                    IMetadataProvider metadataProvider;
-                    _metadataProviders.TryGetValue(connectionInfo, out metadataProvider);
-
-                    queryTab = CreateQueryTab(Path.GetFileName(fn), connectionInfo, metadataProvider);
-                    queryTab.QueryText = File.ReadAllText(fn);
-                    // Modified flag is set during loading because the document 
-                    // "changes" (from nothing to something). So, clear it again.
-                    queryTab.IsDirty = false;
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(this, ex.Message, ex.GetType().Name);
-                    if (queryTab != null)
-                        this.filesDock.RemoveTab(queryTab.Parent as DockContent);
-                    return;
-                }
-
-                // ICSharpCode.TextEditor doesn't have any built-in code folding
-                // strategies, so I've included a simple one. Apparently, the
-                // foldings are not updated automatically, so in this demo the user
-                // cannot add or remove folding regions after loading the file.
-                //--				editor.Document.FoldingManager.FoldingStrategy = new RegionFoldingStrategy();
-                //--				editor.Document.FoldingManager.UpdateFoldings(null, null);
-            }
-        }
-
+        
         private void menuFileClose_Click(object sender, EventArgs e)
         {
             this.filesDock.CloseActiveContent();
@@ -402,7 +263,7 @@ namespace SwqlStudio
         {
             string[] list = e.Data.GetData(DataFormats.FileDrop) as string[];
             if (list != null)
-                OpenFiles(list);
+                this.filesDock.OpenFiles(list);
         }
 
         #endregion
@@ -508,17 +369,7 @@ namespace SwqlStudio
 
         private void menuFileTabPage_Click(object sender, EventArgs e)
         {
-            var tab = this.filesDock.ActiveConnectionTab;
-            if (tab != null)
-            {
-                var connection = tab.ConnectionInfo;
-                var swql = this.filesDock.ActiveQueryTab.QueryText;
-                AddTextToEditor(swql, connection);
-            }
-            else
-            {
-                AddNewQueryTab();
-            }
+            this.filesDock.CreateTabFromPrevious();
         }
 
         private void enableAutocompleteToolStripMenuItem_CheckedChanged(object sender, EventArgs e)
